Refactoring des repertoire. Correction de bugs diverts sur les DAO Recupération du last id sur les DAO Ajout de RaptorDAORecord  qui permet de update juste les valeur que l'on change Gestion des timestamp pour les fonction de date dans l'insert en BDD

+++ www/clean.php
@@ -1,8 +1,11 @@
 <?php 
-	require_once ('../core/common/path.php');
+	require_once ('../vendor/core/common/path.php');
+
+	ini_set('display_errors', 1);
+	ini_set('log_errors', 1);
+	error_reporting(E_ALL);
 	
 	function deleteItem ($path) {
-		
 		if (is_dir ($path)) {
 			foreach (scandir($path) as $item) {
 				if ($item == '.' || $item == '..') {
@@ -10,18 +13,27 @@
 				}
 				deleteItem ($path.'/'.$item);
 			}
+			echo '<span style="color: blue;" >rmdir : </span><span style="color: grey;" >'.$path.'</span><br />';
 			rmdir($path);
 		} else {
+			echo '<span style="color: blue;" >unlink : </span><span style="color: grey;" >'.$path.'</span><br />';
 			unlink($path);
 		}
 	}
 	
 	
 	$d = dir (TEMP_PATH);
-	$listPath = array ();
-	while (false !== ($entry = $d->read())) {
-		if ($entry != '..' && $entry != '.') {
-			deleteItem (TEMP_PATH.$entry);
+	if ($d) {
+		$listPath = array ();
+		while (false !== ($entry = $d->read())) {
+			if ($entry != '..' && $entry != '.') {
+				deleteItem (TEMP_PATH.$entry);
+			}
 		}
+		$d->close();
+
+		echo '<br /><strong>cleaned : </strong><span style="color: green;" >OK</span>';
+	} else {
+		echo '<br /><strong>cleaned : </strong><span style="color: red;" >KO</span>';
 	}
-	$d->close();
+
